Combined UserTest and LaughButton Test
