--- pytests/CreateUserTest.py
@@ -1,46 +1,0 @@
-# Generated by Selenium IDE
-import pytest
-import time
-import json
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-from selenium.webdriver.common.action_chains import ActionChains
-from selenium.webdriver.support import expected_conditions
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
-from selenium.webdriver.chrome.options import Options
-
-class TestCreateUserTest():
-  def setup_method(self, method):
-    options = Options()
-    options.add_argument("--headless")
-    options.add_argument("window-size=1400,1500")
-    options.add_argument("--disable-gpu")
-    options.add_argument("--no-sandbox")
-    options.add_argument("start-maximized")
-    options.add_argument("enable-automation")
-    options.add_argument("--disable-infobars")
-    options.add_argument("--disable-dev-shm-usage")
-    self.driver = webdriver.Chrome(options=options)
-    self.vars = {}
-  
-  def teardown_method(self, method):
-    self.driver.quit()
-  def test_1CreateUserTest(self):
-    self.driver.get("http://127.0.0.1:5000/auth/login?next=%2F")
-    self.driver.set_window_size(1205, 1077)
-    self.driver.find_element(By.ID, "username").send_keys("admin-user")
-    self.driver.find_element(By.ID, "password").send_keys("ABC123")
-    self.driver.find_element(By.LINK_TEXT, "Click to Register!").click()
-    self.driver.find_element(By.ID, "username").click()
-    self.driver.find_element(By.ID, "username").send_keys("Joshua")
-    self.driver.find_element(By.ID, "email").click()
-    self.driver.find_element(By.ID, "email").send_keys("themanjamesgreen@gmail.com")
-    self.driver.find_element(By.ID, "password").click()
-    self.driver.find_element(By.ID, "password").click()
-    self.driver.find_element(By.ID, "password").send_keys("Test1238")
-    self.driver.find_element(By.ID, "password2").click()
-    self.driver.find_element(By.ID, "password2").send_keys("Test1238")
-    self.driver.find_element(By.ID, "submit").click()
-
